requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
faiss-cpu==1.7.4
openai==1.3.0
gunicorn==21.2.0
